Adds make target to manage Archlinux packages
It adds a target to generate the list of installed packages (also on
source control) and second one to install the packages from this list.

Unfortunately I don't have time to investigate how to install
automatically ALSO the AUR packages. I'll leave this for a later point,
at the moment there are other more important priority.